requirements.txt
colorama==0.4.6
greenlet==3.0.3
iniconfig==2.0.0
Jinja2==3.1.4
mariadb==1.1.10
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
pillow==10.4.0
playwright==1.47.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
typing_extensions==4.12.2
